Create mapper.kt
This is a way to run all of the mapscript in a directory and in its children directories (not recursively, just the first folder hierarchy level and the root directory) based on a regex over the addCsv directive in mapscripts. Beside that, there is a log to help resume if some trouble happened during the last execution, not losing everything done already. This script is not done yet